# docs/api-specification.yaml
openapi: 3.0.3
info:
  title: Elite Dangerous Mining Data Server API
  description: |
    # Elite Dangerous Mining Data Server v2.0
    
    A comprehensive API server for Elite Dangerous mining data with real-time statistics, market analysis, and trading route optimization.
    
    ## Features
    - **Real-time EDDN Data Processing**: Live data stream from Elite Dangerous Data Network
    - **Market Data Analysis**: Comprehensive commodity pricing and trading routes
    - **Mining Statistics**: Detailed mining location analytics and profitability tracking
    - **Performance Monitoring**: Server and API usage statistics
    - **WebSocket Integration**: Real-time updates for dynamic data
    
    ## Data Sources
    - **EDDN (Elite Dangerous Data Network)**: Real-time game data via ZeroMQ
    - **Inara API**: Community-driven market and station data  
    - **EDSM API**: System coordinates and exploration data
    
    ## Rate Limiting
    - Standard endpoints: 100 requests per 15 minutes per IP
    - Statistics endpoints: 50 requests per 15 minutes per IP
    - WebSocket connections: 10 concurrent connections per IP
    
    ## Caching
    - Market data: 5-30 minutes TTL depending on volatility
    - Statistics: 1-10 minutes TTL depending on endpoint
    - System data: 24 hours TTL (rarely changes)
    
    ## Response Formats
    All endpoints support JSON by default. Some endpoints also support CSV export via `?format=csv` parameter.
    
  version: "2.0.0"
  contact:
...
servers:
...
paths:
...
  /api/stats:
    get:
      tags:
        - Statistics
      summary: Global server statistics
      description: |
        Returns comprehensive global statistics including:
        - Total systems, stations, and commodities tracked
        - Mining location statistics
        - EDDN message processing statistics
        - API usage metrics
        - WebSocket connection statistics
        - Server performance indicators
      parameters:
        - name: includeHistory
          in: query
          description: Include historical trend data
          schema:
            type: boolean
            default: false
        - name: timeRange
          in: query
          description: Time range for historical data
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: "24h"
        - name: format
          in: query
          description: Response format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Global statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GlobalStatistics'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/stats/eddn:
    get:
      tags:
        - Statistics
      summary: EDDN connection and message statistics
      description: |
        Returns detailed EDDN (Elite Dangerous Data Network) statistics including:
        - Connection status and uptime
        - Message processing rates and types
        - Data validation statistics
        - Network performance metrics
      parameters:
        - name: includeMessageBreakdown
          in: query
          description: Include detailed message type breakdown
          schema:
            type: boolean
            default: true
        - name: timeRange
          in: query
          description: Time range for statistics
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d]
            default: "24h"
      responses:
        '200':
          description: EDDN statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EDDNStatistics'
  /api/stats/mining:
    get:
      tags:
        - Statistics
      summary: Mining data statistics and analytics
      description: |
        Returns comprehensive mining statistics including:
        - Mining location counts and types
        - Commodity distribution analysis
        - Profitability metrics
        - Hotspot efficiency data
      parameters:
        - name: commodity
          in: query
          description: Filter by specific commodity
          schema:
            type: string
        - name: system
          in: query
          description: Filter by system name
          schema:
            type: string
        - name: includeEfficiency
          in: query
          description: Include efficiency analysis
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: Mining statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MiningStatistics'
  /api/stats/api-usage:
    get:
      tags:
        - Statistics
      summary: API usage statistics and performance metrics
      description: |
        Returns API usage statistics including:
        - Request counts by endpoint
        - Response time analytics
        - Error rate tracking
        - Geographic usage distribution
      parameters:
        - name: endpoint
          in: query
          description: Filter by specific endpoint
          schema:
            type: string
        - name: timeRange
          in: query
          description: Time range for usage data
          schema:
            type: string
            enum: [1h, 6h, 24h, 7d, 30d]
            default: "24h"
        - name: includePerformance
          in: query
          description: Include performance metrics
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: API usage statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIUsageStatistics'
...
  /api/market/routes:
    get:
      tags:
        - Market Data
      summary: Calculate optimal trading routes
      description: |
        Calculates optimal trading routes based on:
        - Ship specifications (jump range, cargo capacity)
        - Available credits for trading
        - System security preferences
        - Profit optimization algorithms
      parameters:
        - name: startSystem
          in: query
          required: true
          description: Starting system name
          schema:
            type: string
        - name: jumpRange
          in: query
          description: Ship jump range in light years
          schema:
            type: number
            minimum: 5
            maximum: 100
            default: 20
        - name: cargoCapacity
          in: query
          description: Ship cargo capacity in tons
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: shipSize
          in: query
          description: Ship size for landing pad requirements
          schema:
            type: string
            enum: [small, medium, large]
            default: medium
        - name: credits
          in: query
          description: Available credits for trading
          schema:
            type: integer
            minimum: 1000
            default: 1000000
        - name: maxHops
          in: query
          description: Maximum number of route hops
          schema:
            type: integer
            minimum: 1
            maximum: 10
            default: 3
        - name: minProfitPerTon
          in: query
          description: Minimum profit per ton threshold
          schema:
            type: integer
            minimum: 0
            default: 500
        - name: maxRouteTime
          in: query
          description: Maximum route time in minutes
          schema:
            type: integer
            minimum: 1
            maximum: 1440
            default: 60
        - name: securityLevel
          in: query
          description: Preferred system security level
          schema:
            type: string
            enum: [high, medium, low, any]
            default: any
        - name: stationType
          in: query
          description: Preferred station type
          schema:
            type: string
            enum: [planetary, orbital, any]
            default: any
        - name: includeRisk
          in: query
          description: Include risk assessment
          schema:
            type: boolean
            default: true
        - name: includeVisualization
          in: query
          description: Include route visualization data
          schema:
            type: boolean
            default: true
        - name: sortBy
          in: query
          description: Sort routes by metric
          schema:
            type: string
            enum: [totalProfit, profitPerTon, profitPerHour]
            default: totalProfit
        - name: limit
          in: query
          description: Maximum number of routes to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Trading routes calculated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TradingRoutes'
        '400':
          description: Missing required parameters or invalid values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Start system not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: EDSM service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    GlobalStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            systems:
              type: object
              properties:
                total:
                  type: integer
                withStations:
                  type: integer
                withMiningData:
                  type: integer
                averagePopulation:
                  type: number
            stations:
              type: object
              properties:
                total:
                  type: integer
                withMarketData:
                  type: integer
                withShipyard:
                  type: integer
                withOutfitting:
                  type: integer
            commodities:
              type: object
              properties:
                total:
                  type: integer
                tracked:
                  type: integer
                withPriceData:
                  type: integer
            mining:
              type: object
              properties:
                locations:
                  type: integer
                hotspots:
                  type: integer
                uniqueMaterials:
                  type: integer
            eddn:
              type: object
              properties:
                messagesProcessed:
                  type: integer
                uptime:
                  type: number
                connectionStatus:
                  type: string
            api:
              type: object
              properties:
                totalRequests:
                  type: integer
                averageResponseTime:
                  type: number
                errorRate:
                  type: number
            websocket:
              type: object
              properties:
                activeConnections:
                  type: integer
                totalConnections:
                  type: integer
                messagesDelivered:
                  type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    EDDNStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            connection:
              type: object
              properties:
                status:
                  type: string
                  enum: [connected, disconnected, reconnecting]
                uptime:
                  type: integer
                  description: Connection uptime in milliseconds
                reconnections:
                  type: integer
                lastReconnect:
                  type: string
                  format: date-time
            messages:
              type: object
              properties:
                total:
                  type: integer
                perMinute:
                  type: number
                byType:
                  type: object
                  additionalProperties:
                    type: integer
                lastProcessed:
                  type: string
                  format: date-time
            performance:
              type: object
              properties:
                averageLatency:
                  type: number
                  description: Average processing latency in milliseconds
                throughput:
                  type: number
                  description: Messages per second
                queueSize:
                  type: integer
                memoryUsage:
                  type: integer
            validation:
              type: object
              properties:
                validMessages:
                  type: integer
                invalidMessages:
                  type: integer
                schemaErrors:
                  type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    MiningStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            locations:
              type: object
              properties:
                total:
                  type: integer
                byType:
                  type: object
                  properties:
                    asteroidBelts:
                      type: integer
                    planetRings:
                      type: integer
                    hotspots:
                      type: integer
                uniqueSystems:
                  type: integer
            commodities:
              type: object
              properties:
                distribution:
                  type: object
                  additionalProperties:
                    type: number
                    description: Percentage distribution
                topProfitable:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      avgProfitPerHour:
                        type: number
                      locations:
                        type: integer
            profitability:
              type: object
              properties:
                topLocations:
                  type: array
                  items:
                    type: object
                    properties:
                      system:
                        type: string
                      body:
                        type: string
                      material:
                        type: string
                      profitPerHour:
                        type: number
                        description: Credits per hour
                averageProfitPerHour:
                  type: number
                totalMinersActive:
                  type: integer
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    APIUsageStatistics:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            requests:
              type: object
              properties:
                total:
                  type: integer
                byEndpoint:
                  type: object
                  additionalProperties:
                    type: integer
                byMethod:
                  type: object
                  additionalProperties:
                    type: integer
                perMinute:
                  type: number
            performance:
              type: object
              properties:
                averageResponseTime:
                  type: number
                  description: Average response time in milliseconds
                slowestEndpoints:
                  type: array
                  items:
                    type: object
                    properties:
                      endpoint:
                        type: string
                      avgResponseTime:
                        type: number
                p95ResponseTime:
                  type: number
                p99ResponseTime:
                  type: number
            errors:
              type: object
              properties:
                total:
                  type: integer
                rate:
                  type: number
                  description: Error rate percentage
                byStatus:
                  type: object
                  additionalProperties:
                    type: integer
            geographic:
              type: object
              properties:
                topCountries:
                  type: array
                  items:
                    type: object
                    properties:
                      country:
                        type: string
                      requests:
                        type: integer
                      percentage:
                        type: number
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
...
    TradingRoutes:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            routes:
              type: array
              items:
                $ref: '#/components/schemas/TradingRoute'
            startSystem:
              type: object
              properties:
                name:
                  type: string
                coordinates:
                  $ref: '#/components/schemas/Coordinates'
            searchParameters:
              type: object
              properties:
                jumpRange:
                  type: number
                cargoCapacity:
                  type: integer
                credits:
                  type: integer
                maxHops:
                  type: integer
            summary:
              type: object
              properties:
                totalRoutes:
                  type: integer
                bestTotalProfit:
                  type: integer
                bestProfitPerTon:
                  type: integer
                averageRouteTime:
                  type: number
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
    TradingRoute:
      type: object
      properties:
        routeId:
          type: string
        hops:
          type: array
          items:
            $ref: '#/components/schemas/RouteHop'
        totalProfit:
          type: integer
        profitPerTon:
          type: integer
        profitPerHour:
          type: integer
        estimatedTime:
          type: number
          description: Total route time in minutes
        totalDistance:
          type: number
          description: Total route distance in light years
        riskAssessment:
          type: object
          properties:
            overall:
              type: string
              enum: [low, medium, high]
            factors:
              type: array
              items:
                type: string
        visualization:
          type: object
          nullable: true
          properties:
            waypoints:
              type: array
              items:
                $ref: '#/components/schemas/Coordinates'
            routeLine:
              type: array
              items:
                $ref: '#/components/schemas/Coordinates'
    RouteHop:
      type: object
      properties:
        hopNumber:
          type: integer
        action:
          type: string
          enum: [buy, sell, travel]
        system:
          type: string
        station:
          type: string
          nullable: true
        commodity:
          type: string
          nullable: true
        quantity:
          type: integer
          nullable: true
        price:
          type: integer
          nullable: true
        profit:
          type: integer
          nullable: true
        distance:
          type: number
          description: Distance to this hop in light years
        estimatedTime:
          type: number
          description: Time for this hop in minutes
        coordinates:
          $ref: '#/components/schemas/Coordinates'
...
    Coordinates:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        z:
          type: number
    ResponseMetadata:
      type: object
      properties:
        source:
          type: string
          enum: [cache, database, api]
        processingTime:
          type: number
          description: Processing time in milliseconds
        timestamp:
          type: string
          format: date-time
        cacheStatus:
          type: string
          enum: [hit, miss, disabled]
        dataFreshness:
          type: string
          enum: [fresh, moderate, stale]
        version:
          type: string
    Error:
      type: object
      properties:
        success:
          type: boolean
          default: false
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          nullable: true
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for enhanced rate limits
security:
  - ApiKeyAuth: []
...